refactor: eliminate hex parsing code duplication
- Create shared src/lib/block-utils.js with centralized block utilities
- Move CATEGORY_BLOCKS, parseHexFormatToSelectedBlocks, generateBlockOrder,
  and initializeBlockSelectionFromOnlyBlocks to shared location
- Update imports in make-toolbox-xml.js, block-display-modal.jsx,
  and block-display.js reducers
- Remove duplicate function definitions and unused imports
- All tests passing, no lint errors

Fixes code duplication between make-toolbox-xml.js and
block-display-modal.jsx
